updated scripts for ros2-foxy automatic update. Started attempt to fix linkage of Ros2-foxy libraries. Figuring out that exist some error inside LinkFiles() inside LinuxToolChain.cs

+++ Scripts/replace_references.py
@@ -1,11 +1,11 @@
 from subprocess import check_output
-import os, sys, shutil
+import os, sys, shutil, re
  
 ros2 = '/home/vilkun/work/turtlebot3-UE/Plugins/rclUE/Source/ThirdParty/ros2lib' # sys.argv[1]
 ros2Libs = ros2 + '/lib/'
 
 def RunCommandForEveryLib(commandArgsList):
-    print('Running command:', ' '.join(commandArgsList))
+    print('>', ' '.join(commandArgsList))
     for dirpath,subdirs,files in os.walk(ros2):
         for file in files:
             if file.endswith('.so') or '.so.' in file:
@@ -24,18 +24,55 @@ def ReplaceSonameWithFileRemove(soname, fileName):
     libFileNameOld = ros2Libs + fileName
 
     # leave only one file, starting from *.so.*, but rename it after all to *.so
-    if os.path.exists(libFileNameOld):
+    if libFileNameNew != libFileNameOld and os.path.exists(libFileNameOld):
         if os.path.exists(libFileNameNew):
             os.remove(libFileNameNew)
 
         os.rename(libFileNameOld, libFileNameNew)
 
-    print('Patching soname and all reference to it:', soname)
-    os.system('patchelf --set-soname ' + soname + ' ' + libFileNameNew)
+    command = 'patchelf --set-soname ' + soname + ' ' + libFileNameNew
+    print('>', command)
+    os.system(command)
     RunCommandForEveryLib(['patchelf', '--replace-needed', fileName, soname])
 
-# ReplaceSonameWithFileRemove('libfastcdr.so', 'libfastcdr.so.1')
-# ReplaceSonameWithFileRemove('libfastrtps.so', 'libfastrtps.so.2')
+versionMarker = '.so.'
+libsReplacements = dict()
+for dirpath,subdirs,files in os.walk(ros2):
+        for file in files:
+            if file.endswith('.so') or versionMarker in file:
+                fullName = os.path.join(dirpath, file)
+                lddInfoRaw = os.popen('ldd ' + fullName).read()
+                
+                for rawInfoLine in lddInfoRaw.split('\n'):
+                    if versionMarker in rawInfoLine and 'not found' in rawInfoLine:
+                        libNameVersioning = rawInfoLine.split('=>')[0].lstrip().rstrip()
+                        libName = libNameVersioning.split(versionMarker)[0] + '.so'
+                        potentialFullName = ros2Libs + libName
+
+                        if os.path.exists(potentialFullName):
+                            libsReplacements[libNameVersioning] = libName
+                        else:
+                            print('[Error] Failed to find potential lib:', potentialFullName)
+
+
+libsReplacements = dict(sorted(libsReplacements.items())) 
+#[print(libNameVersioning, libName) for libNameVersioning, libName in libsReplacements.items()]
+for libNameVersioning, libName in libsReplacements.items():
+    ReplaceSonameWithFileRemove(libName, libNameVersioning)
+
+RunCommandForEveryLib(['patchelf', '--force-rpath', '--set-rpath', r'${ORIGIN}:/lib/x86_64-linux-gnu/'])
+#/usr/local/lib:/usr/local/lib64
+
+
+#print('Patching rpath for every lib...')
+#RunCommandForEveryLib(['patchelf', '--force-rpath', '--set-rpath', r'${ORIGIN}'])
+
+#ReplaceSonameWithFileRemove('libfastcdr.so', 'libfastcdr.so.1')
+#ReplaceSonameWithFileRemove('libfastrtps.so', 'libfastrtps.so.2')
+
+
+#ReplaceSonameWithFileRemove('librmw_dds_common__rosidl_typesupport_fastrtps_cpp.so', 'librmw_dds_common__rosidl_typesupport_fastrtps_cpp.so')
+
 # ReplaceSonameWithFileRemove('libconsole_bridge.so', 'libconsole_bridge.so.1.0')
 # ReplaceSonameWithFileRemove('libyaml-cpp.so', 'libyaml-cpp.so.0.6')
 # ReplaceSonameWithFileRemove('liburdfdom_model.so', 'liburdfdom_model.so.1.0')
@@ -48,11 +85,9 @@ def ReplaceSonameWithFileRemove(soname, fileName):
 # ReplaceSonameWithFileRemove('libOgrePaging.so', 'libOgrePaging.so.1.12.1 ')
 # ReplaceSonameWithFileRemove('libOgreOverlay.so', 'libOgreOverlay.so.1.12.1 ')
 
-# print('Patching rpath for every lib...')
-# RunCommandForEveryLib(['patchelf', '--force-rpath', '--set-rpath', r'${ORIGIN}'])
 
 
-print('Removing references to not found libs...')
+# print('Removing references to not found libs...')
 # RunCommandForEveryLib(['patchelf', '--remove-needed', 'libnddsc.so'])
 # RunCommandForEveryLib(['patchelf', '--remove-needed', 'libnddscore.so'])
 # RunCommandForEveryLib(['patchelf', '--remove-needed', 'libnddscpp.so'])
@@ -63,8 +98,8 @@ print('Removing references to not found libs...')
 # RunCommandForEveryLib(['patchelf', '--remove-needed', 'libbuiltin_interfaces__rosidl_typesupport_connext_cpp.so'])
 # RunCommandForEveryLib(['patchelf', '--remove-needed', 'libOgrePaging.so.1.12.1'])
 # RunCommandForEveryLib(['patchelf', '--remove-needed', 'libOgreOverlay.so.1.12.1'])
-RunCommandForEveryLib(['patchelf', '--remove-needed', 'libbuiltin_interfaces__python.so'])
-RunCommandForEveryLib(['patchelf', '--remove-needed', 'libgeometry_msgs__python.so'])
+# RunCommandForEveryLib(['patchelf', '--remove-needed', 'libbuiltin_interfaces__python.so'])
+# RunCommandForEveryLib(['patchelf', '--remove-needed', 'libgeometry_msgs__python.so'])
